requirements.txt
python-dotenv
pypdf
langchain
langchain_openai
langchain-community
pinecone
langchain_cohere
chromadb 
sentence-transformers
groq
streamlit